+++ CHANGELOG.md
@@ -37,6 +37,7 @@
   - Fix TAESD model type detection  
   - Fix LoRA loader incorrectly reporting errors  
   - Fix hypertile for img2img and inpaint operations  
+  - Fix prompt parser batch size  
 
 ## Update for 2025-06-16